Create i secretly want someone to duct tape me to a ceiling.js
i'm serious, i actually want that. this sort of thing hits the sweet spot of craziness and silliness that really defines me as a person, but without the factor of being too over-the-top to make me uncomfortable with it. it may seem to you like it's a frivolous, unnecessary prank that may take a lot of work, but i honestly want somebody to do that to me; it's the sort of prank that i'd genuienly appreciate. you know what? get me drunk to blackout, and have me wake up the next day duct taped to a ceiling. as long as you get me down safely, i'll forever respect you.

adam from mythbusters tested it and duct tape is definitely strong enough to stick someone to a ceiling and hold them there without too much trouble at all.